frontend_pyqt6/requirements.txt
# PyQt6 GUI Framework
PyQt6>=6.6.0
pyqtgraph>=0.13.0

# HTTP client for API calls
requests>=2.31.0
aiohttp>=3.9.0

# Configuration handling
ruamel.yaml>=0.18.0

# Windows notifications
winrt-Windows.UI.Notifications>=1.0.0; sys_platform == "win32"
win10toast>=0.9; sys_platform == "win32"

# Utilities
loguru>=0.7.0